refactor undersampler init
